modified some column's type of database and implement handle reply exception

+++ src/main/java/top/sdugyf/community/community/controller/QuestionController.java
@@ -18,7 +18,7 @@ public class QuestionController {
 
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id,
+    public String question(@PathVariable(name = "id") Long id,
                            Model model) {
 
         QuestionDTO questionDTO = questionService.getById(id);
