+++ src/main/resources/db/migration/V3__Create_comment_table.sql
@@ -3,7 +3,7 @@ create table comment
     id bigint auto_increment,
     parent_id bigint not null,
     type int not null,
-    commentator int not null,
+    commentator bigint not null,
     gmt_create bigint not null,
     gmt_modified bigint not null,
     like_count bigint ,
